requirements.txt
Django==3.1.4
djangorestframework==3.12.2
djoser==2.0.5
psycopg2-binary==2.8.6
